sound-search-API/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
transformers==4.46.3
torch==2.5.1
torchvision==0.20.1
psycopg2-binary==2.9.10
numpy==1.26.4
pillow==11.0.0
av==13.1.0
aiofiles==24.1.0
